requirements.txt
# FastAPI and async web framework
fastapi==0.116.1
uvicorn[standard]==0.35.0

# Core web dependencies
requests==2.32.4
python-dotenv==1.1.1
aiohttp==3.12.14

# Data validation and serialization
pydantic==2.11.7
pydantic-core==2.33.2

# Firebase integration
firebase-admin==7.0.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1

# Composio SDK for unified API integrations (WORKING VERSION)
composio-core==0.7.20

# Required by Composio
sentry-sdk==2.33.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jsonref==1.1.0
inflection==0.5.1
semver==3.0.4
rich==13.9.4
click==8.2.1
pyperclip==1.9.0
pillow==10.4.0
paramiko==3.5.1
pysher==1.0.8
importlib-metadata==8.7.0

# Google AI/Genai integration (WORKING VERSIONS)
google-genai==1.27.0
google-generativeai==0.3.2
google-cloud-aiplatform==1.105.0
google-auth==2.40.3
google-api-core==2.25.1
google-api-python-client==2.176.0
googleapis-common-protos==1.70.0

# Additional Google Cloud dependencies
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-resumable-media==2.7.2
google-crc32c==1.7.1
grpc-google-iam-v1==0.14.2

# HTTP Client with retry capabilities  
httpx==0.28.1
tenacity==8.5.0

# Additional core dependencies
websocket-client==1.8.0
typing-extensions==4.14.1
typing-inspection==0.4.1
anyio==4.9.0
sniffio==1.3.1
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0

# Testing framework
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-mock==3.11.1

# Development tools
black==23.9.1
flake8==6.1.0

# Production server
gunicorn==21.2.0

# Logging
structlog==23.1.0

# Additional FastAPI dependencies
aiofiles==23.2.1
python-multipart==0.0.6
psutil==5.9.5

# JWT Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# Legacy Flask dependencies (for backward compatibility)
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7